+++ includes/fields/media/template.php
@@ -8,8 +8,8 @@
 
 defined( 'ABSPATH' ) || exit;
 
-$field_id           = $this->get_field_id();
-$field_name         = $this->get_field_name();
+$field_id           = $this->get_escaped_field_id();
+$field_name         = $this->get_escaped_field_name();
 $value              = $this->get_escaped_value();
 $button_text        = $this->get_button_text();
 $change_button_text = $this->get_change_button_text();
@@ -19,7 +19,7 @@ $remove_button_text = $this->get_remove_button_text();
 
 <div
 	class="media-field-wrapper"
-	data-field-id="<?php echo esc_attr( $field_id ); ?>"
+	data-field-id="<?php echo $field_id; ?>"
 	data-has-media="<?php echo $value ? 'true' : 'false'; ?>"
 	data-media-type="<?php echo esc_attr( $this->get_media_type() ); ?>"
 	data-media-library-title="<?php echo esc_attr( $this->get_media_library_title() ); ?>"
@@ -30,14 +30,13 @@ $remove_button_text = $this->get_remove_button_text();
 >
 	<input
 		type="hidden"
-		id="<?php echo esc_attr( $field_id ); ?>"
-		name="<?php echo esc_attr( $field_name ); ?>"
-		value="<?php echo esc_attr( $value ); ?>"
+		id="<?php echo $field_id; ?>"
+		name="<?php echo $field_name; ?>"
+		value="<?php echo $value; ?>"
 	/>
 
 	<div
-		class="media-preview"
-		style="<?php echo $value ? '' : 'display:none;'; ?>"
+		class="media-preview<?php echo $value ? '' : ' hidden'; ?>"
 	></div>
 
 	<button
@@ -49,8 +48,7 @@ $remove_button_text = $this->get_remove_button_text();
 
 	<button
 		type="button"
-		class="button media-remove-button"
-		style="<?php echo $value ? '' : 'display:none;'; ?>"
+		class="button media-remove-button<?php echo $value ? '' : ' hidden'; ?>"
 		title="<?php echo esc_attr( $remove_button_text ); ?>"
 	>
 		<span class="dashicons dashicons-no"></span>